fastapi-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
azure-search-documents==11.4.0
azure-core==1.30.0
openai>=1.52.0
python-dotenv==1.0.1
pydantic==2.9.2
pypdf==4.3.1
